Adding undertow ordered cipher configuration
Transfer of the code previously in the jhipster/generator-jhipster

It will force the order of the cipher define in the properties file.
The purpose of this is to achieve perfect forward secrecy for SSL (More infos) in combination of the cipher list in the comments SSL section in the spring boot properties file.